[skip ci] hacky way of detecting incoming connections on a serial port and spawning a new console + terminal to handle them
