+++ kernel/console/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+authors = ["Kevin Boos <kevinaboos@gmail.com>"]
+name = "console"
+description = "Creation and management of virtual consoles/terminals"
+version = "0.1.0"
+build = "../../build.rs"
+
+[dependencies]
+spin = "0.9.0"
+bare-io = { version = "0.2.1", features = [ "alloc" ] }
+log = "0.4.8"
+
+[dependencies.async_channel]
+path = "../async_channel"
+
+[dependencies.serial_port]
+path = "../serial_port"
+
+[dependencies.io]
+path = "../io"
+
+[dependencies.text_terminal]
+path = "../text_terminal"
+
+[dependencies.task]
+path = "../task"
+
+[dependencies.spawn]
+path = "../spawn"
+
+[dependencies.irq_safety]
+git = "https://github.com/theseus-os/irq_safety"
+
+[dependencies.mpmc]
+path = "../../libs/mpmc"
+
+[lib]
+crate-type = ["rlib"]
