+++ kernel/serial_port/Cargo.toml
@@ -6,7 +6,6 @@ version = "0.1.0"
 build = "../../build.rs"
 
 [dependencies]
-log = "0.4.8"
 spin = "0.9.0"
 bare-io = { version = "0.2.1", features = [ "alloc" ] }
 
@@ -16,8 +15,15 @@ path = "../../libs/port_io"
 [dependencies.irq_safety]
 git = "https://github.com/theseus-os/irq_safety"
 
+# Dependencies below here are temporary, for console creation testing.
 [dependencies.mpmc]
 path = "../../libs/mpmc"
+
+[dependencies.io]
+path = "../io"
+
+[dependencies.async_channel]
+path = "../async_channel"
 
 [lib]
 crate-type = ["rlib"]
