+++ kernel/text_terminal/Cargo.toml
@@ -9,6 +9,7 @@ build = "../../build.rs"
 bitflags = "1.1.0"
 bare-io = { version = "0.2.1", features = [ "alloc", "nightly" ] }
 unicode-width = "0.1.8"
+vte = "0.10.1"
 
 [dependencies.log]
 version = "0.4.8"
